+++ iceoryx_posh/source/roudi/memory/mempool_collection_memory_block.cpp
@@ -39,7 +39,7 @@ MemPoolCollectionMemoryBlock::~MemPoolCollectionMemoryBlock() noexcept
 
 uint64_t MemPoolCollectionMemoryBlock::size() const noexcept
 {
-    return cxx::align(static_cast<uint64_t>(sizeof(mepoo::MemoryManager)), mepoo::MemPool::CHUNK_MEMORY_ALIGNMENT)
+    return cxx::align(sizeof(mepoo::MemoryManager), mepoo::MemPool::CHUNK_MEMORY_ALIGNMENT)
            + mepoo::MemoryManager::requiredFullMemorySize(m_memPoolConfig);
 }
 
@@ -47,7 +47,7 @@ uint64_t MemPoolCollectionMemoryBlock::alignment() const noexcept
 {
     // algorithm::align doesn't like constexpr values
     auto memPoolAlignment = mepoo::MemPool::CHUNK_MEMORY_ALIGNMENT;
-    return algorithm::max(static_cast<uint64_t>(alignof(mepoo::MemoryManager)), memPoolAlignment);
+    return algorithm::max(alignof(mepoo::MemoryManager), memPoolAlignment);
 }
 
 void MemPoolCollectionMemoryBlock::onMemoryAvailable(cxx::not_null<void*> memory) noexcept
